Create class DFNMesh
